+++ src/test/java/qna/domain/DeleteHistoriesTest.java
@@ -1,0 +1,37 @@
+package qna.domain;
+
+import org.assertj.core.api.Assertions;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.Test;
+import qna.CannotDeleteException;
+
+import java.time.LocalDateTime;
+import java.util.Arrays;
+import java.util.List;
+
+import static org.assertj.core.api.Assertions.assertThat;
+
+
+class DeleteHistoriesTest {
+    private Question question;
+    private Answer answer;
+
+    @BeforeEach
+    public void setUp() throws Exception {
+        question = new Question(1L, "title1", "contents1").writeBy(UserTest.JAVAJIGI);
+        answer = new Answer(11L, UserTest.JAVAJIGI, QuestionTest.Q1, "Answers Contents1");
+        question.addAnswer(answer);
+    }
+
+    @Test
+    void delete_성공() throws CannotDeleteException {
+        List<DeleteHistory> result = Arrays.asList(
+                new DeleteHistory(ContentType.QUESTION, question.getId(), question.getWriter(), LocalDateTime.now()),
+                new DeleteHistory(ContentType.ANSWER, answer.getId(), answer.getWriter(), LocalDateTime.now()));
+
+        DeleteHistories deleteHistories = new DeleteHistories(question.delete(UserTest.JAVAJIGI, 1L)
+                , question.getAnswers().deleteAnswers(UserTest.JAVAJIGI));
+
+        assertThat(deleteHistories.getDeleteHistories().size()).isEqualTo(2);
+    }
+}
